requirements.txt
openai>=1.40.0
openenv-core>=0.2.2
pydantic>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
